Refactoring and javadoc

+++ src/statistics/DataPreprocessing.java
@@ -7,17 +7,29 @@ import models.PaasProfile;
 
 public interface DataPreprocessing {
 
-	public Map<String, Long> evalRevision(List<PaasProfile> profiles);
+	/**
+	 * Evaluates the age of Revision of a PaasProfile in Days
+	 * 
+	 * @param profiles
+	 * @return A Map with keys: latest, oldest, mean and their value in Days.
+	 */
+	public Map<String, Double> evalRevision(List<PaasProfile> profiles);
 
 	public Map<String, Long> evalStatus(List<PaasProfile> profiles);
 
+	/**
+	 * Evaluates the age of StatusSince of a PaasProfile in Days
+	 * 
+	 * @param profiles
+	 * @return A Map with keys: latest, oldest, mean and their value in Days.
+	 */
 	public Map<String, Double> evalStatusSince(List<PaasProfile> profiles);
 
 	public Map<String, Long> evalType(List<PaasProfile> profiles);
-	
+
 	public Map<String, Long> evalQos(List<PaasProfile> profiles);
-	
-	public Map<String,Long> evalPlatform(List<PaasProfile> profiles);
+
+	public Map<String, Long> evalPlatform(List<PaasProfile> profiles);
 
 	public Map<String, Long> evalHosting(List<PaasProfile> profiles);
 
@@ -25,16 +37,26 @@ public interface DataPreprocessing {
 
 	public Map<String, Long> evalScaling(List<PaasProfile> profiles);
 
+	/**
+	 * Evaluates the Runtime of a PaasProfile. A Profile might have multiple
+	 * languages and a language might have multiple versions. For every Version
+	 * of a language there is a key in the Map returned with the frequency as
+	 * its value. The DataFormat of the key is ("language"|"version"). A version
+	 * consists either of numbers of wildcards ("*").
+	 * 
+	 * @param profiles
+	 * @return Map with language and version as a key and its frequency as value
+	 */
 	public Map<String, Long> evalRuntimes(List<PaasProfile> profiles);
-	
+
 	public Map<String, Long> evalMiddleware(List<PaasProfile> profiles);
-	
+
 	public Map<String, Long> evalFrameworks(List<PaasProfile> profiles);
-	
+
 	public Map<String, Long> evalServices(List<PaasProfile> profiles);
-	
+
 	public Map<String, Long> evalExtensible(List<PaasProfile> profiles);
-	
+
 	public Map<String, Long> evalInfrastructures(List<PaasProfile> profiles);
 
 }
